backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
requests
